+++ src/constraints/table.jl
@@ -156,10 +156,8 @@ function update_table(com::CoM, constraint::TableConstraint)
         vidx = indices[local_vidx]
         var = variables[vidx]
         clear_mask(current)
-        nremoved = num_removed(var, backtrack_idx)
-        # don't use for fixed variables or if number of values is smaller than removed values
-        if nremoved < nvalues(var) && nvalues(var) != 1
-            for value in view_removed_values(variables[vidx], nremoved)
+        if num_removed(var) < nvalues(var) && !isfixed(var)
+            for value in view_removed_values(var)
                 add_to_mask(current, supports[com, vidx, local_vidx, value])
             end
             invert_mask(current)
@@ -169,7 +167,7 @@ function update_table(com::CoM, constraint::TableConstraint)
                 add_to_mask(current, supports[com, vidx, local_vidx, value])
             end
         end
-       
+
         intersect_with_mask(current)
         is_empty(current) && break
     end
@@ -246,8 +244,6 @@ function prune_constraint!(
         !feasible && break
     end
     empty!(constraint.changed_vars)
-    # full mask for `single_reverse_pruning_constraint!``
-    full_mask(current)
     return feasible
 end
 
@@ -294,7 +290,6 @@ function still_feasible(
     end
     feasible = intersect_with_mask_feasible(current)
 
-    full_mask(current)
     return feasible
 end
 
@@ -439,7 +434,6 @@ function reverse_pruning_constraint!(
         end
         # otherwise there is nothing to reverse
     end
-    full_mask(current)
     reset_residues!(com, constraint)
     empty!(constraint.changed_vars)
 end
@@ -465,13 +459,23 @@ function restore_pruning_constraint!(
     constraint.changed_vars = 1:length(constraint.std.indices)
     current = constraint.current
     backtrack_id = last(prune_steps)
-    while backtrack_id > length(backtrack_id) || isempty(constraint.backtrack[backtrack_id].words)
+    while backtrack_id > length(constraint.backtrack) || isempty(constraint.backtrack[backtrack_id].words)
         backtrack_id = com.backtrack_vec[backtrack_id].parent_idx
     end
     current.last_ptr = constraint.backtrack[backtrack_id].last_ptr
     current.words = copy(constraint.backtrack[backtrack_id].words)
     current.indices = copy(constraint.backtrack[backtrack_id].indices)
-    full_mask(current)
     reset_residues!(com, constraint)
     empty!(constraint.changed_vars)
+end
+
+function is_solved_constraint(com::CoM,
+    constraint::TableConstraint,
+    fct::MOI.VectorOfVariables,
+    set::TableSetInternal,
+)
+
+    values = CS.value.(com.search_space[constraint.std.indices])
+    table = set.table
+    return findfirst(ri->table[ri,:] == values, 1:size(table)[1]) !== nothing
 end