+++ test/eternity.jl
@@ -35,7 +35,7 @@ end
 
     ncolors = maximum(puzzle[:,2:end])
 
-    m = Model(CSJuMPTestSolver())
+    m = Model(optimizer_with_attributes(CS.Optimizer, "keep_logs"=>true, "logging"=>[]))
     @variable(m, 1 <= p[1:height, 1:width] <= npieces, Int)
     @variable(m, 0 <= pu[1:height, 1:width] <= ncolors, Int)
     @variable(m, 0 <= pr[1:height, 1:width] <= ncolors, Int)
@@ -81,45 +81,11 @@ end
     end
 
     optimize!(m)
+    com = JuMP.backend(m).optimizer.model.inner
+    @test is_solved(com)
 
     status = JuMP.termination_status(m)
     @test status == MOI.OPTIMAL
 
-    # testing that constraints are fulfilled
-    sp = convert.(Int, JuMP.value.(p))
-    spu = convert.(Int, JuMP.value.(pu))
-    spr = convert.(Int, JuMP.value.(pr))
-    spd = convert.(Int, JuMP.value.(pd))
-    spl = convert.(Int, JuMP.value.(pl))
-    @test allunique(sp)
-
-    for i=1:height, j=1:width
-        @test findfirst(ri->rotations[ri,:] == [sp[i,j], spu[i,j], spr[i,j], spd[i,j], spl[i,j]], 1:length(rotations)) !== nothing
-    end
-
-    for j=1:width
-        @test spu[1,j] == 0
-        @test spd[height,j] == 0
-
-        if j != width
-            @test spr[1,j] == spl[1,j+1]
-            @test spr[height,j] == spl[height,j+1]
-        end
-    end
-
-     # right and left
-     for i=1:height
-        @test spl[i,1] == 0
-        @test spr[i,width] == 0
-
-        if i != height
-            @test spd[i,1] == spu[i+1,1]
-            @test spd[i,width] == spu[i+1,width]
-        end
-    end
-
-    for i=1:height-1, j=1:width-1
-        @test spd[i,j] == spu[i+1, j]
-        @test spr[i,j] == spl[i, j+1]
-    end
+    @test general_tree_test(com)
 end