+++ test/maximum_weight_matching.jl
@@ -17,7 +17,7 @@
         ]
         n = 6
 
-        m = Model(CSJuMPTestSolver())
+        m = Model(CSJuMPTestOptimizer())
         @variable(m, x[1:n, 1:n], Bin)
         for i = 1:n, j = 1:n
             if weight_matrix[i, j] == 0 || i > j