+++ test/sudoku.jl
@@ -136,7 +136,7 @@
             0 0 0 0 0 0 8 0 6
         ]
 
-        m = Model(CSJuMPTestSolver())
+        m = Model(CSJuMPTestOptimizer())
         @variable(m, 1 <= x[1:9, 1:9] <= 9, Int)
         # set variables
         for r = 1:9, c = 1:9
@@ -164,7 +164,7 @@
         grid[8, :] = [0 8 0 0 7 3 0 5 0]
         grid[9, :] = [0 0 0 0 0 0 0 0 0]
 
-        m = Model(CSJuMPTestSolver())
+        m = Model(CSJuMPTestOptimizer())
         @variable(m, 1 <= x[1:9, 1:9] <= 9, Int)
         # set variables
         nvars_set = 0
@@ -206,7 +206,7 @@
         grid[8, :] = [0 0 0 0 7 0 0 5 0]
         grid[9, :] = [0 0 1 0 0 0 0 0 0]
 
-        m = Model(CSJuMPTestSolver())
+        m = Model(CSJuMPTestOptimizer())
         @variable(m, 1 <= x[1:9, 1:9] <= 9, Int)
         # set variables
         nvars_set = 0
@@ -265,7 +265,7 @@
         options = CS.combine_options(options)
 
         @test CS.solve!(com, options) == :Solved
-        @test fulfills_sudoku_constr(com_grid)
+        @test is_solved(com)
     end
 
     @testset "Hard fsudoku repo 0-8 Int8 Objective" begin
@@ -311,7 +311,7 @@
         options = CS.combine_options(options)
 
         @test CS.solve!(com, options) == :Solved
-        @test fulfills_sudoku_constr(com_grid)
+        @test is_solved(com)
         @test typeof(com.best_bound) == Int8
         @test typeof(com.best_sol) == Int8
     end
@@ -344,7 +344,7 @@
             @test typeof(com.best_sol) == Int8
             @test JuMP.objective_value(m) == JuMP.value(x[1, 1]) == com.best_sol
             @test JuMP.termination_status(m) == MOI.OPTIMAL
-            @test jump_fulfills_sudoku_constr(JuMP.value.(x))
+            @test is_solved(com)
             c += 1
         end
         # check that actually all 95 problems were tested